修改所有的id类型(long -> int) 新增返回给前端的Result类 新增登录凭证token类

+++ src/main/java/velore/po/BlogTag.java
@@ -22,9 +22,9 @@ import org.springframework.stereotype.Component;
 public class BlogTag {
 
     @TableId(type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
-    private Long blogId;
+    private Integer blogId;
 
-    private Long tagId;
+    private Integer tagId;
 }
